+++ src/components/items/PreviewMosaic.vue
@@ -1,11 +1,11 @@
 <template>
   <a :href="'item/'+item.id+'/'">
-    <Image :url="coverUrl" :title="item.fields[collection.titleField]" :cover=true />
+    <Image :url="coverUrl" :title="item.properties[collection.titleProperty]" :cover=true />
   </a>
 </template>
 
 <script>
-import Image from '@/components/fields/medias/Image.vue'
+import Image from '@/components/properties/medias/Image.vue'
 
 export default {
   components: {
@@ -22,8 +22,8 @@ export default {
       return this.$parent.collection
     },
     coverUrl() {
-      if (this.item.fields[this.collection.coverField]) {
-        return this.item.fields[this.collection.coverField]
+      if (this.item.properties[this.collection.coverProperty]) {
+        return this.item.properties[this.collection.coverProperty]
       } else {
         return this.collection.cover
       }