+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog for OnMyShelf web interface
 
 ## 1.0 Beta 2 (202X-XX-XX)
+- Rename 'fields' to 'properties' for better understanding
 - Various bugfixes
 
 ## 1.0 Beta 1 (2021-10-09)